% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{combo}
\alias{combo}
\title{Combo String}
\usage{
combo(metadata, suffix = NULL)
}
\arguments{
\item{metadata}{STAR Metadata object.}

\item{suffix}{an optional string to paste after the combo, e.g. a file
extension (including the period).}
}
\value{
String with the format \code{"STAR_2019_HKE_5"} followed by an optional
\code{suffix}.
}
\description{
Create a string that combines the Reference Year, Species, and GSA.
}
\note{
When a stock assessment covers multiple GSAs, they are separated by
underscores. An extreme example is \code{"STAR_2019_HKE_8_9_10_11.1_11.2"}.
}
\examples{
\dontrun{

combo(star$Metadata)

combo(star$Metadata, ".xlsx")
}

}
\seealso{
\code{\link{gsa.names}} converts comma-separated GSA codes to full GSA names.

\code{\link{lookup.species}} is a data frame containing GSA codes and names.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
