% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfcmSTAR-package.R
\docType{package}
\name{gfcmSTAR-package}
\alias{gfcmSTAR-package}
\alias{gfcmSTAR}
\title{Tools to Work with GFCM STAR}
\description{
Tools that support the Stock Assessment Results (STAR) framework of the
General Fisheries Commission for the Mediterranean.
}
\details{
\emph{Read Excel template into R:}
\tabular{ll}{
  \code{\link{read.template.v10}} \tab version 1.0
}
\emph{Write R object to CSV files:}
\tabular{ll}{
  \code{\link{write.star}} \tab \code{metadata.csv} and \code{timeseries.csv}
}
}
