% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsa.names.R
\name{gsa.names}
\alias{gsa.names}
\title{Convert GSA Names}
\usage{
gsa.names(gsa)
}
\arguments{
\item{gsa}{comma-separated GSA codes.}
}
\value{
String with the format \code{"1 - Northern Alboran Sea, 2 - Alboran Island"}.
}
\description{
Convert comma-separated GSA codes to a series of full GSA names.
}
\examples{
gsa.names("1,2")

}
\seealso{
\code{\link{lookup.gsa}} is a data frame containing GSA codes and names.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
