% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identical.stars.R
\name{identical.stars}
\alias{identical.stars}
\title{Identical STARs}
\usage{
identical.stars(x, y)
}
\arguments{
\item{x}{STAR template filename or STAR object, a list containing
\code{Metadata} and \code{TimeSeries}.}

\item{y}{another STAR template/object to compare.}
}
\value{
\code{TRUE} or \code{FALSE}, indicating whether the objects are identical,
disregarding the SharePoint and Excel properties.
}
\description{
Check whether two STAR templates/objects are identical. This comparison
excludes the SharePoint and Excel properties \code{SharePoint_Folder},
\code{Excel_Filename}, \code{Person_Modified}, \code{Time_Modified}, and
\code{Time_Imported}.
}
\note{
When importing many STAR templates submitted by stock assessors, it can
happen that two Excel files contain the same stock assessment results. This
function helps to identify such cases.
}
\examples{
\dontrun{

# Are the R objects completely identical?
identical(star1, star2)

# Do the STAR objects contain the same data,
# disregarding SharePoint and Excel properties?
identical.stars(star1, star2)

# Work directly with files
identical.stars("STAR_1.xlsx", "STAR_2.xlsx")
}

}
\seealso{
\code{\link{diff.stars}} shows differences between STAR templates/objects.

\code{\link{identical}} is the underlying function used to compare the two
objects.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
