% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.dir.R
\name{import.dir}
\alias{import.dir}
\title{Import Directory}
\usage{
import.dir(dir, exclude = NULL, ...)
}
\arguments{
\item{dir}{directory name.}

\item{exclude}{filenames to exclude.}

\item{\dots}{passed to \code{read.template}.}
}
\value{
List of STAR objects.
}
\description{
Import all Excel STAR templates from a directory to a list.
}
\examples{
\dontrun{
cluster <- import.dir("STAR")
}

}
\seealso{
\code{\link{read.template}} reads an Excel STAR template into a STAR object.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
