% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.many.templates.R
\name{import.many.templates}
\alias{import.many.templates}
\title{Import Many Templates}
\usage{
import.many.templates(dir, pattern = "\\\\.xlsx?$", exclude = NULL,
  short = TRUE, qc = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{dir}{directory name.}

\item{pattern}{regular expression to select filenames to include.}

\item{exclude}{filenames to exclude.}

\item{short}{whether to show the filename in a short \code{\link{basename}}
format.}

\item{qc}{whether to call \code{\link{qc}} on each filename.}

\item{quiet}{whether to suppress messages.}

\item{\dots}{passed to \code{read.template}.}
}
\value{
A cluster (list) of STAR objects.
}
\description{
Import many Excel STAR templates from a directory into a cluster (list) of
STAR objects.
}
\details{
The default \code{pattern} selects all filenames ending with \file{xls} or
\file{xlsx}.
}
\note{
A vignette demonstrates the use of \code{import.many.templates} to import
STAR templates:
\preformatted{
vignette("import", "gfcmSTAR")
}
}
\examples{
\dontrun{
cluster <- import.many.templates("STAR")
}

}
\seealso{
\code{\link{read.template}} reads an Excel STAR template into a STAR object.

\code{\link{import.many.csv}} imports many STAR CSV files from a directory
tree into a cluster.

\code{\link{peek}} examines a metadata field of STAR objects.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
