% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.gsa.R
\docType{data}
\name{lookup.gsa}
\alias{lookup.gsa}
\title{Geographical Subareas (GSA)}
\format{
Data frame containing 2 columns:
\tabular{ll}{
  \code{Code} \tab GSA code\cr
  \code{Name} \tab GSA name
}
}
\source{
\url{http://www.fao.org/gfcm/data/maps/gsas/en/}
}
\usage{
lookup.gsa
}
\description{
Reference table of GFCM Geographical Subareas (GSA) in FAO Major Fishing Area
37 - Mediterranean and Black Sea.
}
\examples{
lookup.gsa
}
\seealso{
\code{\link{gsa.names}} converts comma-separated GSA codes to GSA full names.

\code{\link{lookup.species}} is a data frame containing species codes and
names.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
