% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.species.R
\docType{data}
\name{lookup.species}
\alias{lookup.species}
\title{Species}
\format{
Data frame containing 3 columns:
\tabular{ll}{
  \code{Alpha}      \tab 3-alpha species code\cr
  \code{Scientific} \tab Scientific name\cr
  \code{English}    \tab English name
}
}
\source{
\url{https://gfcm.sharepoint.com/sites/DCRF/Lists/Ref_Species/AllItems.aspx}
}
\usage{
lookup.species
}
\description{
Reference table of GFCM species.
}
\examples{
head(lookup.species)
}
\seealso{
\code{\link{lookup.gsa}} is a data frame containing GSA codes and names.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
