% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Peek}
\usage{
peek(x, field = "Assessment_ID")
}
\arguments{
\item{x}{STAR object or a cluster (list) of STAR objects. Alternatively,
\code{x} can be an Excel STAR template filename or directory name.}

\item{field}{a metadata field, by default "Assessment_ID".}
}
\value{
Metadata field from STAR object(s).
}
\description{
Examine a metadata field of STAR objects.
}
\note{
This function is convenient to quickly examine a cluster (list) of STAR
objects after importing many STAR templates. See example below.
}
\examples{
\dontrun{

peek(star)
peek(star, "Scientific_Name")
peek(star, "Status_Fref")

peek(star)         # object
peek(cluster)      # cluster
peek("STAR.xlsx")  # file
peek("folder")     # dir
}

}
\seealso{
\code{\link{import.many.templates}} imports many Excel STAR templates from a
directory into a cluster (list).

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
