% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.all.R
\name{qc.all}
\alias{qc.all}
\title{Quality Check}
\usage{
qc.all(file, stop = TRUE)
}
\arguments{
\item{file}{filename of a STAR template.}

\item{stop}{whether to stop if test fails.}
}
\value{
\code{TRUE} if all tests succeed, otherwise an error message (if
\code{stop = TRUE}) or FALSE and a warning message (if \code{stop = FALSE}).
}
\description{
Run all quality checks on a STAR template.
}
\examples{
\dontrun{
qc.all("STAR_2019_HKE_5.xlsx")
}

}
\seealso{
The checks are run in the following order:

\code{\link{qc.xlsx}} checks if file has an \file{xlsx} extension.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
