% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.colnames.R
\name{qc.colnames}
\alias{qc.colnames}
\title{Column Names}
\usage{
qc.colnames(file, stop = TRUE, quiet = FALSE)
}
\arguments{
\item{file}{filename of an Excel STAR template.}

\item{stop}{whether to stop if test fails.}

\item{quiet}{whether to suppress messages.}
}
\value{
\code{TRUE} if test succeeds, otherwise an error message
(if \code{stop = TRUE}) or \code{FALSE} and a warning message
(if \code{stop = FALSE}).
}
\description{
Assert that a STAR template's time series column names are intact.
}
\note{
This function compares the column names found in the STAR template sheet
containing the time series against the expected column names. The expected
column names are those defined inside the \code{\link{set.classes}} function.
}
\examples{
\dontrun{
qc.colnames("STAR_2019_HKE_5.xlsx")
}

}
\seealso{
\code{\link{qc}} runs all \code{qc.*} tests.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
