% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.exists.R
\name{qc.exists}
\alias{qc.exists}
\title{File Exists}
\usage{
qc.exists(file, stop = TRUE, quiet = FALSE)
}
\arguments{
\item{file}{filename of a STAR template.}

\item{stop}{whether to stop if test fails.}

\item{quiet}{whether to suppress messages.}
}
\value{
\code{TRUE} if test succeeds, otherwise an error message
(if \code{stop = TRUE}) or \code{FALSE} and a warning message
(if \code{stop = FALSE}).
}
\description{
Assert that a file exists.
}
\examples{
\dontrun{
qc.exists("STAR_2019_HKE_5.xlsx")
}

}
\seealso{
\code{\link{qc.all}} runs all \code{qc.*} tests.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
