% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.star.R
\name{qc.star}
\alias{qc.star}
\title{STAR Template}
\usage{
qc.star(file, stop = TRUE, quiet = FALSE)
}
\arguments{
\item{file}{filename of a STAR template.}

\item{stop}{whether to stop if test fails.}

\item{quiet}{whether to suppress messages.}
}
\value{
\code{TRUE} if test succeeds, otherwise an error message
(if \code{stop = TRUE}) or \code{FALSE} and a warning message
(if \code{stop = FALSE}).
}
\description{
Assert that a file is a STAR template and contains a \code{Metadata} sheet.
}
\examples{
\dontrun{
qc.star("STAR_2019_HKE_5.xlsx")
}

}
\seealso{
\code{\link{qc}} runs all \code{qc.*} tests.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
