% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.ts.numbers.R
\name{qc.ts.numbers}
\alias{qc.ts.numbers}
\title{Time Series Numbers}
\usage{
qc.ts.numbers(file, short = TRUE, stop = TRUE, quiet = FALSE)
}
\arguments{
\item{file}{filename of an Excel STAR template.}

\item{short}{whether to show the filename in a short \code{\link{basename}}
format.}

\item{stop}{whether to stop if test fails.}

\item{quiet}{whether to suppress messages.}
}
\value{
\code{TRUE} if test succeeds, otherwise an error message
(if \code{stop = TRUE}) or \code{FALSE} and a warning message
(if \code{stop = FALSE}).
}
\description{
Assert that a STAR template's time series are numbers and not strings.
}
\note{
Stock assessors sometimes enter values into STAR templates as strings instead
of numbers. This is usually because of regional settings on different
computers.
}
\examples{
\dontrun{
qc.ts.numbers("STAR_2019_HKE_5.xlsx")
}

}
\seealso{
\code{\link{qc}} runs all \code{qc.*} tests.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
