% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.vpa.R
\name{qc.vpa}
\alias{qc.vpa}
\title{VPA Model Type}
\usage{
qc.vpa(file, short = TRUE, stop = TRUE, quiet = FALSE)
}
\arguments{
\item{file}{filename of an Excel STAR template.}

\item{short}{whether to show the filename in a short \code{\link{basename}}
format.}

\item{stop}{whether to stop if test fails.}

\item{quiet}{whether to suppress messages.}
}
\value{
\code{TRUE} if test succeeds, otherwise an error message
(if \code{stop = TRUE}) or \code{FALSE} and a warning message
(if \code{stop = FALSE}).
}
\description{
Assert that a STAR template's table object \code{VPA_Model} is either
\code{"Yes"} or \code{"No"}.
}
\examples{
\dontrun{
qc.vpa("STAR_2019_HKE_5.xlsx")
}

}
\seealso{
\code{\link{qc}} runs all \code{qc.*} tests.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
