% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.properties.R
\name{read.properties}
\alias{read.properties}
\title{Read Properties}
\usage{
read.properties(file, template = NULL)
}
\arguments{
\item{file}{Excel file containing the SharePoint properties.}

\item{template}{optional filename of an Excel STAR template.}
}
\value{
Data frame with SharePoint properties.
}
\description{
Read SharePoint properties from an Excel file.
}
\note{
To download Excel SharePoint properties:
\enumerate{
\item Starting in Teams, navigate to STAR folder.
\item Select \sQuote{Open in SharePoint}.
\item In SharePoint, select \sQuote{Export to Excel}.
\item Save file with a descriptive filename, such as
      \file{WGSAD_Western_Mediterranean.xlsx}.
}
}
\examples{
\dontrun{

read.properties("WGSAD_Western_Mediterranean.xlsx")

read.properties("WGSAD_Western_Mediterranean.xlsx", "STAR_2019_HKE_5.xlsx")
}

}
\seealso{
\code{\link[XLConnect]{loadWorkbook}} and \code{\link[XLConnect]{readTable}}
are the underlying functions used to read SharePoint properties from an Excel
file.

\code{\link{read.template.v10}} reads an Excel template into a STAR object.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
