% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.star.v10.R
\name{read.star.v10}
\alias{read.star.v10}
\title{Read STAR Template Version 1.0}
\usage{
read.star.v10(file, atype = "Standard", refyear = 2019, repyear = 2021,
  countries = NA)
}
\arguments{
\item{file}{name of Excel file.}

\item{atype}{assessment type, \code{"Standard"} or \code{"Benchmark"}.}

\item{refyear}{reference year.}

\item{repyear}{reporting year.}

\item{countries}{countries, separated by comma and space.}
}
\value{
List containing \code{Metadata} (list) and \code{TimeSeries} (data frame).
}
\description{
Import the results found in an Excel STAR template version 1.0.
}
\examples{
\dontrun{
star <- read.star.v10("STAR_2019_HKE_5.xlsx")
}

}
\seealso{
\code{\link[XLConnect]{loadWorkbook}}, \code{\link[XLConnect]{readTable}},
and \code{\link[XLConnect]{readWorksheet}} are the underlying functions to
read data from Excel workbooks.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
