% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.template.v10.R
\name{read.template.v10}
\alias{read.template.v10}
\title{Read STAR Template Version 1.0}
\usage{
read.template.v10(file, atype = "Standard", refyear = 2019,
  repyear = 2021, countries = NA)
}
\arguments{
\item{file}{name of Excel file.}

\item{atype}{assessment type, \code{"Standard"} or \code{"Benchmark"}.}

\item{refyear}{reference year.}

\item{repyear}{reporting year.}

\item{countries}{countries, separated by comma and space.}
}
\value{
STAR object, a list containing \code{Metadata} (simple list) and
\code{TimeSeries} (data frame).
}
\description{
Read the results found in an Excel STAR template version 1.0 into a STAR
object.
}
\note{
The purpose of storing \code{Metadata} as class
\code{c("simple.list", "list")} is to have it display in a compact and
readable format in the console. It is easy to convert to a normal list or
data frame - see examples below.
}
\examples{
\dontrun{

# Import
star <- read.template.v10("STAR_2019_HKE_5.xlsx")

# Coerce metadata to list or data frame:
star$Metadata                 # simple.list (default)
unclass(star$Metadata)        # list
as.data.frame(star$Metadata)  # data.frame
}

}
\seealso{
\code{\link[XLConnect]{loadWorkbook}} and \code{\link[XLConnect]{readTable}}
are the underlying functions used to read data from Excel STAR templates.

\code{\link{write.star}} writes a STAR object to CSV files.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
