% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report Files Successfully Imported}
\usage{
report(cluster, cluster.ok, qc.vector)
}
\arguments{
\item{cluster}{a list of STAR objects, including ones that resulted in import
\code{"try-error"}.}

\item{cluster.ok}{a list of STAR objects that have been selected and prepared
for final import into the STAR database.}

\item{qc.vector}{a logical vector, output from \code{qc}.}
}
\value{
List containing file \code{Count} and \code{Filenames}.
}
\description{
Show a summary report after importing and quality checking a directory of
STAR templates.
}
\examples{
\dontrun{

# Directory names
uploads <- "~/StockAssessmentResults/uploads/2021"
star.dir <- file.path(uploads, "WGSASP_General")
prop.file <- file.path(uploads, "properties/WGSASP_General.xlsx")

# Import
prop <- read.properties(prop.file)
cluster <- import.many.templates(star.dir, prop=prop)

# Select valid STAR objects to import into STAR database
cluster.ok <- cluster[sapply(cluster, class) != "try-error"]
cluster.ok$"star_template.xlsx" <- NULL

# Quality check
qc.vector <- qc(star.dir)

# Report
report(cluster, cluster.ok, qc.vector)
}

}
\seealso{
\code{\link{import.many.templates}} imports many Excel STAR templates from a
directory into a cluster (list).

\code{\link{qc}} runs all QC tests on an Excel STAR template or a directory
containing Excel STAR templates.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
