% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.version.R
\name{template.version}
\alias{template.version}
\title{Template Version}
\usage{
template.version(file)
}
\arguments{
\item{file}{Excel STAR template, with the version number specified on the
first sheet.}
}
\value{
String containing version number.
}
\description{
Detect STAR template version of an Excel file.
}
\examples{
\dontrun{
template.version("STAR_2019_HKE_5.xlsx")
}

}
\seealso{
\code{\link{read.template}} reads an Excel template into a STAR object.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
