% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.report.R
\name{write.report}
\alias{write.report}
\title{Write Report}
\usage{
write.report(rep, file, nchar = 23)
}
\arguments{
\item{rep}{a \code{\link{report}} object.}

\item{file}{a filename to write to.}

\item{nchar}{maximum number of characters in error messages.}
}
\description{
Write a report object to a text file.
}
\details{
When displaying a \code{report} object, error messages from filenames that
failed to import are truncated to 23 characters by default. This makes the
report easy to read from the console or a text file. To show full error
messages, pass a large value, such as \code{nchar = 999}.
}
\examples{
\dontrun{
rep <- report(cluster, cluster.ok, qc.vector)
write.report(rep, "report.txt")
}

}
\seealso{
\code{\link{report}} reports which files were successfully imported.

\code{\link{writeLines}} is the underlying function used to write the file.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
