% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.star.csv.R
\name{write.star.csv}
\alias{write.star.csv}
\title{Write STAR CSV}
\usage{
write.star.csv(star, dir = TRUE, topdir = NULL, mfile = "metadata.csv",
  tfile = "timeseries.csv", quote = TRUE, row.names = FALSE,
  fileEncoding = "UTF-8", dos = TRUE, quiet = FALSE, force = FALSE,
  ...)
}
\arguments{
\item{star}{STAR object, a list containing \code{Metadata} and
\code{TimeSeries}.}

\item{dir}{an optional directory to write the CSV files into, or a logical
value. See details below.}

\item{topdir}{an optional directory containing \code{dir}. See details below.}

\item{mfile}{a filename for the metadata.}

\item{tfile}{a filename for the time series.}

\item{quote}{whether to quote strings.}

\item{row.names}{whether to include row names.}

\item{fileEncoding}{character encoding for output file.}

\item{dos}{whether to ensure the resulting CSV files have Dos line endings
(CRLF).}

\item{quiet}{whether to suppress messages.}

\item{force}{whether to overwrite existing files.}

\item{\dots}{passed to \code{write.csv}.}
}
\value{
\code{TRUE} if CSV files were created, otherwise \code{FALSE}.
}
\description{
Write a STAR object to CSV files \file{metadata.csv} and
\file{timeseries.csv}.
}
\details{
The special value \code{dir = TRUE} can be used as shorthand for
\code{dir = star$Metadata$Assessment_ID}, which may be a suitable directory
name.

The special value \code{dir = FALSE} is equivalent to \code{dir = "."} which
will write the CSV files into the current directory.

The \code{topdir} argument can be used to organize the output from multiple
STAR objects in one top directory. For example, \preformatted{
hake_4 <- read.template("STAR_2019_HKE_4.xlsx")
hake_5 <- read.template("STAR_2019_HKE_5.xlsx")
write.star.csv(hake_4, topdir="csv")
write.star.csv(hake_5, topdir="csv")} will produce
four files in the following directory structure: \preformatted{
csv/
    STAR_2019_HKE_4/
        metadata.csv
        timeseries.csv
    STAR_2019_HKE_5/
        metadata.csv
        timeseries.csv}
}
\note{
A vignette demonstrates the use of \code{write.star.csv} to export a list of
STAR objects to CSV files:
\preformatted{
vignette("export", "gfcmSTAR")
}
}
\examples{
\dontrun{

write.star.csv(star)

write.star.csv(star, dir="hake_5")

write.star.csv(star, topdir="csv")
}

}
\seealso{
\code{\link{write.csv}} is the underlying function used to write a table to a
CSV file.

\code{\link{read.template}} reads an Excel STAR template into a STAR object.

\code{\link{read.star.csv}} reads a pair of STAR CSV files into a STAR
object.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
