% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.star.R
\name{write.star}
\alias{write.star}
\title{Write STAR Object to Files}
\usage{
write.star(star, dir = NULL, mfile = "metadata.csv",
  tfile = "timeseries.csv", quote = TRUE, row.names = FALSE,
  fileEncoding = "UTF-8", dos = TRUE, ...)
}
\arguments{
\item{star}{a STAR object, list containing \code{Metadata} and
\code{TimeSeries}.}

\item{dir}{an optional directory name.}

\item{mfile}{a filename for the metadata.}

\item{tfile}{a filename for the time series.}

\item{quote}{whether to quote strings.}

\item{row.names}{whether to include row names.}

\item{fileEncoding}{character encoding for output file.}

\item{dos}{whether to ensure the resulting CSV files have Dos line endings.}

\item{\dots}{passed to \code{write.csv}.}
}
\value{
No return value, called for side effects.
}
\description{
Write a STAR object to CSV files.
}
\seealso{
\code{\link{write.csv}} is the underlying function used to write a table to a
file.

\code{\link{read.star.v10}} reads from an Excel template into a STAR object.
}
