% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{combo}
\alias{combo}
\title{Combo String}
\usage{
combo(star, ...)
}
\arguments{
\item{star}{STAR object containing \code{Metadata}.}

\item{\dots}{optional strings to \code{\link{paste0}} after the combo. See
examples below.}
}
\value{
String with the format \code{"STAR_2019_HKE_5"} followed by optional strings
passed by user.
}
\description{
Create a string that combines Reference Year, Species Code, and GSA.
}
\note{
When a stock assessment covers multiple GSAs, they are combined in a compact
format. For example GSA \code{"8,9,10,11.1,11.2"} becomes \code{"891011"}.
}
\examples{
\dontrun{

combo(star)

combo(star, ".xlsx")

combo(star, "_", star$Metadata$Assessment_Method)
}

}
\seealso{
\code{\link{lookup.species}} is a data frame containing species codes and
names.

\code{\link{gsa.names}} converts comma-separated GSA codes to full GSA names.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
