% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{combo}
\alias{combo}
\title{Combo String}
\usage{
combo(metadata, ...)
}
\arguments{
\item{metadata}{STAR Metadata object.}

\item{\dots}{optional strings to \code{\link{paste0}} after the combo, e.g. a
file extension (including the period) or assessment method.}
}
\value{
String with the format \code{"STAR_2019_HKE_5"} followed by optional strings
passed as \dots.
}
\description{
Create a string that combines Reference Year, Species Code, and GSA.
}
\note{
When a stock assessment covers multiple GSAs, they are separated by
underscores. An extreme example is \code{"STAR_2019_HKE_8_9_10_11.1_11.2"}.
}
\examples{
\dontrun{

combo(star$Metadata)

combo(star$Metadata, ".xlsx")

combo(star$Metadata, "_", star$Metadata$Assessment_Method)
}

}
\seealso{
\code{\link{lookup.species}} is a data frame containing species codes and names.

\code{\link{gsa.names}} converts comma-separated GSA codes to full GSA names.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
