% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.many.csv.R
\name{export.many.csv}
\alias{export.many.csv}
\title{Export Many CSV Files}
\usage{
export.many.csv(cluster, topdir, force = FALSE, quiet = TRUE, ...)
}
\arguments{
\item{cluster}{list of STAR objects.}

\item{topdir}{top directory that will contain STAR subdirectories.}

\item{force}{whether to overwrite existing files.}

\item{quiet}{whether to suppress messages.}

\item{\dots}{passed to \code{write.star.csv}.}
}
\value{
\code{TRUE} if CSV files were created, otherwise \code{FALSE}.
}
\description{
Export a cluster (list) of STAR objects to CSV files \file{metadata.csv} and
\file{timeseries.csv}.
}
\note{
A vignette demonstrates the use of \code{export.many.csv} to export a cluster
(list) of STAR objects to CSV files:
\preformatted{
vignette("export", "gfcmSTAR")
}
}
\examples{
\dontrun{
export.many.csv(cluster, topdir="working_group")
}

}
\seealso{
\code{\link{write.star.csv}} writes a STAR object to CSV files.

\code{\link{import.many.csv}} imports many STAR CSV files from a directory
tree into a cluster.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
