% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma2period.R, R/gfcmSTAR-internal.R,
%   R/readTableLogical.R, R/readTableTranspose.R
\name{comma2period}
\alias{comma2period}
\alias{gfcmSTAR-internal}
\alias{readTableLogical}
\alias{readTableTranspose}
\title{Various gfcmSTAR Helper Functions}
\usage{
comma2period(x)

readTableLogical(...)

readTableTranspose(..., rownames = 1)
}
\description{
These functions are called by user-level functions. The functionality is
documented in the user-level functions.
}
\seealso{
\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
\keyword{internal}
