% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma2period.R, R/gfcmSTAR-internal.R,
%   R/gsa.names.R, R/readTableLogical.R, R/readTableTranspose.R, R/write.star.R
\name{comma2period}
\alias{comma2period}
\alias{gfcmSTAR-internal}
\alias{gsa.names}
\alias{readTableLogical}
\alias{readTableTranspose}
\alias{write.star}
\title{Various gfcmSTAR Helper Functions}
\usage{
comma2period(x)

gsa.names(gsa)

readTableLogical(...)

readTableTranspose(..., rownames = 1)

write.star(star, mfile = "metadata.csv", tfile = "timeseries.csv", ...)
}
\description{
These functions are called by user-level functions. The functionality is
documented in the user-level functions.
}
\seealso{
\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
\keyword{internal}
