% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma2period.R, R/gfcmSTAR-internal.R,
%   R/readTable.logical.R, R/readTable.transpose.R
\name{comma2period}
\alias{comma2period}
\alias{gfcmSTAR-internal}
\alias{readTable.logical}
\alias{readTable.transpose}
\title{Various gfcmSTAR Helper Functions}
\usage{
comma2period(x)

readTable.logical(...)

readTable.transpose(..., rownames = 1)
}
\description{
These functions are called by user-level functions. The functionality is
documented in the user-level functions.
}
\seealso{
\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
\keyword{internal}
