% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Start_Case.R, R/basic.time.R, R/comma2period.R,
%   R/gfcmSTAR-internal.R, R/readTable.logical.R, R/readTable.transpose.R,
%   R/report.R
\name{Start_Case}
\alias{Start_Case}
\alias{basic.time}
\alias{comma2period}
\alias{gfcmSTAR-internal}
\alias{readTable.logical}
\alias{readTable.transpose}
\alias{cat.report}
\alias{print.report}
\alias{write.report}
\title{Various gfcmSTAR Helper Functions}
\usage{
Start_Case(x)

basic.time(x)

comma2period(x)

readTable.logical(...)

readTable.transpose(..., rownames = 1, cols = 1)

cat.report(x, nchar = 23)

\method{print}{report}(x, nchar = 23, ...)

write.report(x, file, nchar = 23)
}
\description{
These functions are called by user-level functions. The functionality is
documented in the user-level functions.
}
\seealso{
\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
\keyword{internal}
