% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfcmSTAR-package.R
\docType{package}
\name{gfcmSTAR-package}
\alias{gfcmSTAR-package}
\alias{gfcmSTAR}
\title{Tools to Work with GFCM STAR}
\description{
Tools that support the Stock Assessment Results (STAR) framework of the
General Fisheries Commission for the Mediterranean.
}
