% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfcmSTAR-package.R
\docType{package}
\name{gfcmSTAR-package}
\alias{gfcmSTAR-package}
\alias{gfcmSTAR}
\title{Tools to Work with GFCM STAR}
\description{
Tools that support the Stock Assessment Results (STAR) framework of the
General Fisheries Commission for the Mediterranean.
}
\details{
\emph{Import:}
\tabular{ll}{
  \code{\link{read.template.v10}} \tab read STAR template v1.0\cr
  \code{\link{read.template.v10}} \tab read SharePoint properties
}
\emph{Export:}
\tabular{ll}{
  \code{\link{write.star}} \tab write STAR object to CSV files
}
\emph{Lookup tables:}
\tabular{ll}{
  \code{\link{lookup.gsa}} \tab geographical subareas (GSA)
}
}
