% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfcmSTAR-package.R
\docType{package}
\name{gfcmSTAR-package}
\alias{gfcmSTAR-package}
\alias{gfcmSTAR}
\title{Tools to Work with GFCM Stock Assessment Results (STAR)}
\description{
Tools that support the Stock Assessment Results (STAR) framework of the
General Fisheries Commission for the Mediterranean.
}
\details{
\emph{Import:}
\tabular{ll}{
  \code{\link{read.template}}     \tab read STAR template\cr
  \code{\link{read.template.v10}} \tab read STAR template version 1.0\cr
  \code{\link{read.properties}}   \tab read SharePoint properties\cr
  \code{\link{template.version}}  \tab detect STAR template version
}
\emph{Export:}
\tabular{ll}{
  \code{\link{write.star}} \tab write STAR object to CSV files
}
\emph{Format strings:}
\tabular{ll}{
  \code{\link{combo}}     \tab combine reference year, species, and GSA\cr
  \code{\link{gsa.names}} \tab convert GSA codes to full GSA names
}
\emph{Lookup tables:}
\tabular{ll}{
  \code{\link{lookup.gsa}}     \tab geographical subareas (GSA)\cr
  \code{\link{lookup.species}} \tab species
}
\emph{Quality control:}
\tabular{ll}{
  \code{\link{qc.all}} \tab run all quality checks\cr
  \code{\link{qc.xlsx}} \tab file extension
}
}
