% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identical.stars.R
\name{identical.stars}
\alias{identical.stars}
\title{Identical STARs}
\usage{
identical.stars(a, b)
}
\arguments{
\item{a}{STAR object, a list containing \code{Metadata} and
\code{TimeSeries}.}

\item{b}{another STAR object.}
}
\value{
\code{TRUE} or \code{FALSE}, indicating whether the objects are identical,
disregarding the time when they were imported.
}
\description{
Check whether two STAR objects are identical. This comparison excludes the
\code{Time_Imported} data field.
}
\note{
When importing many STAR templates submitted by stock assessors, it can
happen that two Excel files contain the same data. This function helps to
identify such cases.
}
\examples{
\dontrun{
identical.star(star1, star2)
}

}
\seealso{
\code{\link{identical}} is the underlying base function used to compare the
two objects.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
