% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Peek}
\usage{
peek(x, field)
}
\arguments{
\item{x}{STAR object or a cluster (list) of STAR objects.}

\item{field}{a metadata field, such as "Assessment_ID".}
}
\value{
Metadata field from STAR object(s).
}
\description{
Examine a metadata field of STAR objects.
}
\note{
This function is convenient to quickly examine a cluster (list) of STAR
objects after importing many STAR templates. See example below.
}
\examples{
\dontrun{

peek(star, "Assessment_ID")

peek(cluster, "Assessment_ID")
}

}
\seealso{
\code{\link{import.many.templates}} imports many Excel STAR templates from a
directory into a cluster (list).

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
