% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc}
\alias{qc}
\title{Quality Check}
\usage{
qc(x, short = TRUE, stop = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{filename of an Excel STAR template or a directory containing Excel
STAR templates.}

\item{short}{whether to show filenames in a short \code{\link{basename}}
format.}

\item{stop}{whether to stop if any test fails.}

\item{quiet}{whether to suppress messages.}
}
\value{
Logical \code{TRUE} or \code{FALSE} indicating whether all tests succeeded,
or a logical vector if \code{x} is a directory.
}
\description{
Run all quality checks on one or more Excel STAR templates.
}
\details{
If \code{x} is a directory, \code{stop} is set to \code{FALSE} unless the
user passes an explicit \code{stop = TRUE}.
}
\examples{
\dontrun{

# Filename
qc("STAR_2019_HKE_5.xlsx")

# Directory
ok <- qc("working_group")
ok[!ok]
}

}
\seealso{
The checks are run in the following order:

\code{\link{qc.exists}} checks if file exists.

\code{\link{qc.xlsx}} checks if file extension is \file{xlsx}.

\code{\link{qc.star}} checks if file is a STAR template.

\code{\link{qc.vpa}} checks if \code{VPA_Model} is \code{Yes} or \code{No}.

\code{\link{qc.ts.names}} checks if time series column names are intact.

\code{\link{qc.ts.numbers}} checks if time series are numbers and not
strings.

\code{\link{report}} reports files successfully imported.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
