% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.star.csv.R
\name{read.star.csv}
\alias{read.star.csv}
\title{Read STAR CSV}
\usage{
read.star.csv(dir)
}
\arguments{
\item{dir}{directory name.}
}
\value{
STAR object, a list containing \code{Metadata} (simple list) and
\code{TimeSeries} (data frame).
}
\description{
Read STAR CSV files \file{metadata.csv} and \file{timeseries.csv} into a STAR
object.
}
\examples{
\dontrun{
read.star.csv("STAR_2019_HKE_5")
}

}
\seealso{
\code{\link{read.csv}} is the underlying base function used to read the CSV
files.

\code{\link{read.template}} reads an Excel STAR template into a STAR object.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
