% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.template.R
\name{read.template}
\alias{read.template}
\title{Read STAR Template}
\usage{
read.template(file, ...)
}
\arguments{
\item{file}{name of Excel file.}

\item{\dots}{passed to version-specific functions, such as
\code{read.template.v10}.}
}
\value{
STAR object, a list containing \code{Metadata} (simple list) and
\code{TimeSeries} (data frame).
}
\description{
Read the results found in an Excel STAR template into a STAR object.
}
\note{
The user can either use the general \code{read.template} function or
version-specific functions, such as read.template.v10, to the same effect.

The Excel STAR templates have a version number that consists of three
numbers: \emph{major.minor.patch}. The \dfn{major} version is incremented
when the STAR database design needs to be changed, the \dfn{minor} version
when the R import function needs to be changed, and the \dfn{patch} version
when changes are backwards compatible.

Detailed documentation of the functionality of \code{read.template} is found
in the version-specific help pages, listed below.
}
\examples{
\dontrun{
star <- read.template("STAR_2019_HKE_5.xlsx")
}

}
\seealso{
\code{\link{read.template.v10}} reads an Excel template of version 1.0 into a
STAR object.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
