% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.template.v10.R
\name{read.template.v10}
\alias{read.template.v10}
\title{Read STAR Template Version 1.0}
\usage{
read.template.v10(file, atype = "Standard", refyear = 2019,
  repyear = 2021, countries = NA, prop = NULL, suffix = "",
  quiet = TRUE)
}
\arguments{
\item{file}{name of Excel file.}

\item{atype}{assessment type, \code{"Standard"} or \code{"Benchmark"}.}

\item{refyear}{reference year.}

\item{repyear}{reporting year.}

\item{countries}{countries, separated by comma and space.}

\item{prop}{SharePoint properties from \code{read.properties}.}

\item{suffix}{optional string passed to \code{combo} to construct a unique
\code{Assessment_ID} field.}

\item{quiet}{whether to supress messages.}
}
\value{
STAR object, a list containing \code{Metadata} (simple list) and
\code{TimeSeries} (data frame).
}
\description{
Read the results found in an Excel STAR template version 1.0 into a STAR
object.
}
\examples{
\dontrun{

# Import
star <- read.template.v10("STAR_2019_HKE_5.xlsx")
}

}
\seealso{
\code{\link{read.template}} calls \code{read.template.v10} when the Excel
STAR template is of version 1.0.

\code{\link[XLConnect]{loadWorkbook}} and \code{\link[XLConnect]{readTable}}
are the underlying functions used to read data from Excel STAR templates.

\code{\link{read.properties}} reads SharePoint properties from an Excel file.

\code{\link{combo}} is used to construct the \code{Assessment_ID} metadata
field.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
