% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report Files Successfully Imported}
\usage{
report(cluster, cluster.ok, qc.vector)
}
\arguments{
\item{cluster}{a list of STAR objects, including ones that resulted in import
\code{"try-error"}.}

\item{cluster.ok}{a list of STAR objects that have been selected and prepared
for final import into the STAR database.}

\item{qc.vector}{a logical vector, output from \code{qc}.}
}
\value{
Nested list with top elements \code{Count} and \code{Filenames}.
}
\description{
Show a summary report after importing and quality checking a directory of
STAR templates.
}
\note{
The full set of Excel templates included in the report consists of all the
STAR objects in \code{cluster}. They can be divided into three categories:
\itemize{
\item Imported. The STAR objects found in \code{cluster.ok}.
\item Failed. Files that \code{read.template} could not import.
\item Removed. Files that were manually removed by the administrator, because
      they were duplicate templates, empty templates, etc.
}

The report also includes a count of files that had errors detected by the
\code{qc} suite of quality checks. The QC issues are independent of whether
\code{read.template} was able to import or not. The purpose of the QC is to
raise a flag where stock assessors may have made a mistake when submitting
their STAR templates.

In addition to file counts, the report also includes the specific filenames
that had errors, were removed, or had QC issues. This list is useful for STAR
administrators to bring into focus any STAR templates that were not
successfully imported.
}
\section{Print Method}{


When displaying a \code{report} object, error messages from filenames that
failed to import are truncated to 23 characters by default. This makes the
report easy to read from the console. To show full error messages, pass a
large value of \code{nchar} to the \code{print} method:

\preformatted{rep <- report(cluster, cluster.ok, qc.vector)
print(rep, nchar=999)}
}

\examples{
\dontrun{

# Directory names
uploads <- "~/StockAssessmentResults/uploads/2021"
star.dir <- file.path(uploads, "WGSASP_General")
prop.file <- file.path(uploads, "properties/WGSASP_General.xlsx")

# Import
prop <- read.properties(prop.file)
cluster <- import.many.templates(star.dir, prop=prop)

# Select valid STAR objects to import into STAR database
cluster.ok <- cluster[sapply(cluster, class) != "try-error"]
cluster.ok$"star_template.xlsx" <- NULL

# Quality check
qc.vector <- qc(star.dir)

# Report
report(cluster, cluster.ok, qc.vector)
}

}
\seealso{
\code{\link{import.many.templates}} imports many Excel STAR templates from a
directory into a cluster (list).

\code{\link{qc}} runs all QC tests on an Excel STAR template or a directory
containing Excel STAR templates.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
