% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.star.R
\name{write.star}
\alias{write.star}
\title{Write STAR Object to Files}
\usage{
write.star(x, mfile = "metadata.csv", tfile = "timeseries.csv", ...)
}
\arguments{
\item{x}{a STAR object, list containing \code{Metadata} and
\code{TimeSeries}.}

\item{mfile}{a filename for the metadata.}

\item{tfile}{a filename for the time series.}

\item{\dots}{passed to \code{write.csv}.}
}
\description{
Write a STAR object to files.
}
