% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.star.R
\name{write.star}
\alias{write.star}
\title{Write STAR Object to Files}
\usage{
write.star(star, dir = TRUE, mfile = "metadata.csv",
  tfile = "timeseries.csv", quote = TRUE, row.names = FALSE,
  fileEncoding = "UTF-8", dos = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{star}{STAR object, a list containing \code{Metadata} and
\code{TimeSeries}.}

\item{dir}{an optional directory name to write the CSV files into, or a
logical value. See details below.}

\item{mfile}{a filename for the metadata.}

\item{tfile}{a filename for the time series.}

\item{quote}{whether to quote strings.}

\item{row.names}{whether to include row names.}

\item{fileEncoding}{character encoding for output file.}

\item{dos}{whether to ensure the resulting CSV files have Dos line endings
(CRLF).}

\item{quiet}{whether to suppress messages.}

\item{\dots}{passed to \code{write.csv}.}
}
\value{
No return value, called for side effects.
}
\description{
Write a STAR object to CSV files.
}
\details{
The special value \code{dir = TRUE} can be used as shorthand for the prefix
of the original Excel filename. For example, if
\code{star$Metadata$Excel_Filename} is \code{"STAR_2019_HKE_5.xlsx"}, then a
directory called \file{STAR_2019_HKE_5} will be created and used.

The special value \code{dir = FALSE} can be used as shorthand for
\code{dir = "."} to write into the current directory.
}
\seealso{
\code{\link{write.csv}} is the underlying function used to write a table to a
CSV file.

\code{\link{read.template.v10}} reads an Excel template into a STAR object.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
