% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.star.R
\name{write.star}
\alias{write.star}
\title{Write STAR Object to Files}
\usage{
write.star(star, dir = TRUE, topdir = NULL, mfile = "metadata.csv",
  tfile = "timeseries.csv", quote = TRUE, row.names = FALSE,
  fileEncoding = "UTF-8", dos = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{star}{STAR object, a list containing \code{Metadata} and
\code{TimeSeries}.}

\item{dir}{an optional directory to write the CSV files into, or a logical
value. See details below.}

\item{topdir}{an optional directory containing \code{dir}. See details below.}

\item{mfile}{a filename for the metadata.}

\item{tfile}{a filename for the time series.}

\item{quote}{whether to quote strings.}

\item{row.names}{whether to include row names.}

\item{fileEncoding}{character encoding for output file.}

\item{dos}{whether to ensure the resulting CSV files have Dos line endings
(CRLF).}

\item{quiet}{whether to suppress messages.}

\item{\dots}{passed to \code{write.csv}.}
}
\value{
No return value, called for side effects.
}
\description{
Write a STAR object to CSV files.
}
\details{
The special value \code{dir = TRUE} can be used as shorthand for
\code{combo(star$Metadata)}, which can produce a suitable directory name.

The special value \code{dir = FALSE} is equivalent to \code{dir = "."} which
will write the CSV files into the current directory.

The \code{topdir} argument can be used to organize the output from multiple
STAR objects in one top directory. For example, \preformatted{
write.star(read.template("STAR_2019_HKE_4.xlsx"), topdir="csv")
write.star(read.template("STAR_2019_HKE_5.xlsx"), topdir="csv")} will produce
four files in the following directory structure: \preformatted{
csv/
    STAR_2019_HKE_4/
        metadata.csv
        timeseries.csv
    STAR_2019_HKE_5/
        metadata.csv
        timeseries.csv}
}
\examples{
\dontrun{

write.star(star)

write.star(star, dir="hake_5")

write.star(star, topdir="csv")
}

}
\seealso{
\code{\link{write.csv}} is the underlying function used to write a table to a
CSV file.

\code{\link{read.template}} reads an Excel STAR template into a STAR object.

\code{\link{gfcmSTAR-package}} gives an overview of the package.
}
